fixes: code review

+++ resources/views/ssh/services/database/mysql-family/rollback-networking.blade.php
@@ -1,1 +1,5 @@
-sudo rm -f {{ $dropIn }}
+if sudo test -s {{ $dropIn }}.vito.bak; then
+    sudo cp {{ $dropIn }}.vito.bak {{ $dropIn }}
+else
+    sudo rm -f {{ $dropIn }}
+fi
